Emily's first pass
My first pass through this. I made a lot of the changes to this to mirror the other content that was published with the same steps. I think we might want to enhance the introduction a bit. Why/when do people use DML operations? what problem are they trying to solve when they use then, etc?